Feature ETP-451: Fixing storybook
