+++ packages/ComponentLibrary/src/components/ProfileModal/ToggleButton/index.tsx
@@ -4,11 +4,7 @@ import { toggleContainerStyles, toggleButtonStyles } from './styles';
 import { ToggleSectionsProps } from './types';
 import { theme } from '../../../theme';
 
-const ToggleSections: React.FC<ToggleSectionsProps> = ({
-  sections,
-  currentSection,
-  onToggle,
-}) => {
+const ToggleSections: React.FC<ToggleSectionsProps> = ({ sections, currentSection, onToggle }) => {
   return (
     <div style={toggleContainerStyles}>
       {sections.map(({ id, label, icon }) => {
@@ -18,9 +14,7 @@ const ToggleSections: React.FC<ToggleSectionsProps> = ({
             key={id}
             style={{
               ...toggleButtonStyles,
-              backgroundColor: isActive
-                ? theme.palette.baselineColor.neutral[0]
-                : '',
+              backgroundColor: isActive ? theme.palette.baselineColor.neutral[0] : '',
             }}
             onClick={() => onToggle(id)}
             startIcon={isActive ? icon : null}>
