+++ packages/storybook/src/stories/Components/Nav/Nav.stories.tsx
@@ -78,6 +78,8 @@ export const DefaultNav: Story = {
           roles={[]}
           selectedRole={null}
           selectedWarehouse={null}
+          saveAsDefault={false}
+          onSaveAsDefaultChange={() => {}}
         />
       </div>
     </Nav>
